add snapshot tests for symbols & rename test folders to __test__
