+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ui",
+  "name": "ultimate-tic-tac-react",
   "version": "0.1.0",
   "homepage": "http://maracuja-juice.github.io/ultimate-tic-tac-react",
   "private": true,
@@ -17,12 +17,10 @@
   "scripts": {
     "start": "react-scripts-ts start",
     "build": "react-scripts-ts build",
-    "predeploy": "yarn build",
-    "deploy": "gh-pages -d build",
     "test": "react-scripts-ts test --env=jsdom --coverage",
+    "test:watch": "react-scripts-ts test --watch",
     "eject": "react-scripts-ts eject",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "storybook": "start-storybook -p 6006"
   },
   "jest": {
     "collectCoverageFrom": [
