+++ src/components/Board/SmallBoard/SmallBoard.story.tsx
