+++ src/components/Board/SmallBoard/SmallBoard.test.tsx
