+++ src/components/Board/SmallBoard/SmallBoard.tsx
