+++ src/components/Board/SmallBoard/smallboard.css
