+++ src/components/symbols/__test__/XSymbol.test.tsx
