+++ src/state/sagas/__test__/boardCalculationSaga.test.ts
@@ -4,7 +4,7 @@ import { getBoards } from '../../selectors/AppStateSelectors';
 import { select } from 'redux-saga/effects';
 import { CALCULATE_BOARD_VALUE, SET_BOARD_VALUE } from '../../board/boardActions';
 import { Player } from '../../AppState';
-import unfinishedBoardMock from './unfinishedBoardMock';
+import unfinishedBoardMock from '../../../__mocks__/unfinishedBoardMock';
 
 describe( 'boardCalculationSaga', () => {
     it( 'should dispatch the set board value for the board with the Cross Player', () => {
